+++ sunpy/util/net.py
@@ -176,12 +176,12 @@ def url_exists(url, timeout=2):
 
     Parameters
     ----------
-    url: str
+    url: `str`
         A string containing a URL
 
     Returns
     -------
-    value: bool
+    value: `bool`
 
     Examples
     --------
@@ -199,3 +199,20 @@ def url_exists(url, timeout=2):
         return False
     else:
         return True
+
+
+def is_online():
+    """
+    Checks whether an internet connection is available.
+
+    Returns
+    -------
+    value: `bool`
+
+    Examples
+    --------
+    >>> from sunpy.util.net import is_online
+    >>> is_online()
+    True
+    """
+    return url_exists('http://www.google.com')